### 2023-04-19 #5377 @emdupre
👋 Hi @lucasmiranda42, and thanks for your submission to JOSS !

I see that you've already made a few suggestions for potential reviewers, thank you ! I'll look through these suggestions and send out invitations soon.
