### 2023-04-20 eef21ee
Fixed bug in filter_conditions()
