### 2023-04-21 #5377 @emdupre
👋 Hi @edeno, @cellistigs, @ezhang94,

Would you be willing to review _DeepOF: a Python package for supervised and unsupervised pattern recognition in mice motion tracking data_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/mlfpm/deepof) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.05377/joss.05377/10.21105.joss.05377.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you for considering !
