### 2023-04-21 #5394 @editorialbot
[REVIEW]: DeepOF: a Python package for supervised and unsupervised pattern recognition in mice motion tracking data
**Submitting author:** @lucasmiranda42 (<a href="http://orcid.org/0000-0002-5484-6744">Lucas Miranda</a>)
**Repository:** https://github.com/mlfpm/deepof
**Branch with paper.md** (empty if default branch): JOSS_paper
**Version:** JOSS
