### 2023-04-26 f3cfa95
Fixed bug in filter_conditions()
