+++ pyproject.toml
@@ -43,6 +43,8 @@ joblib = "^1.2.0"
 h5py = "^3.7.0"
 tables = "^3.8.0"
 catboost = "^1.1.1"
+tornado = "6.1"
+jupyter-client = "7.3.2"
 
 [tool.poetry.group.dev.dependencies]
 black = "^22.8.0"
