### 2023-04-26 eda9204
Fixed bug in plot_gantt()
