### 2023-04-26 24cf800
Fixed bug in filter_condition()
