### 2023-04-27 #5394 @lucasmiranda42
Hi all! And thank you very much for reviewing our package! 😃

Quick question: are we allowed to push to the paper branch while reviews are ongoing? We're pushing some refactoring to master, and I was wondering if I should push here too. Nothing major is changing (neither API nor anything depicted in the tutorials).

Best, and thanks a lot once again!
Lucas
