### 2023-04-27 0f1e1d7
Added tests for condition filtering
