### 2023-04-28 562ff16
Refactored statannotations usage to enable the comparison of more than two categories per experimental condition in all plotting functions
