### 2023-05-01 9ca42f6
Refactored tutorials and added missing tests
