+++ CONTRIBUTING.md
@@ -1,0 +1,86 @@
+# Contributing to DeepOF
+
+We appreciate your interest in contributing to DeepOF! This guide details how to contribute to the DeepOF project in a way that is efficient for everyone.
+
+## Table of Contents
+
+- [Code of Conduct](#code-of-conduct)
+- [Getting Started](#getting-started)
+- [How Can I Contribute?](#how-can-i-contribute)
+- [Style Guide](#style-guide)
+- [Pull Requests](#pull-requests)
+- [Issue Tracking](#issue-tracking)
+
+## Code of Conduct
+
+All contributors are expected to adhere to the project's [Code of Conduct](CODE_OF_CONDUCT.md). Please read the document before contributing. 
+
+## Getting Started
+
+1. Fork the project repository and clone your fork:
+   
+    ```
+    git clone git@github.com:mlfpm/deepof.git
+    ```
+
+2. Create a new branch for your changes:
+   
+    ```
+    git checkout -b name-of-your-branch
+    ```
+
+3. Make your changes and commit them:
+   
+    ```
+    git commit -m "Detailed commit message"
+    ```
+
+4. Push your changes to your fork:
+   
+    ```
+    git push origin name-of-your-branch
+    ```
+
+5. Create a pull request from your branch to the DeepOF main branch.
+
+## How Can I Contribute?
+
+There are many ways you can contribute, beyond writing code. If you can write, draw, design, or organize, we would love your help. Here are some ways to contribute:
+
+- Improve documentation
+- Fix bugs or add new features
+- Design UI or graphics
+- Write tutorials or blog posts
+- Review code submissions
+- Test the application and report issues
+
+## Style Guide
+
+We use the [Black](https://black.readthedocs.io/en/stable/) code formatter for this project. Please ensure your contributions adhere to this standard before submitting a pull request. You can check and format your code using Black:
+
+    ```
+    black your_file.py
+    ```
+
+Black will automatically format your code to meet the style guide.
+
+## Pull Requests
+
+Pull requests are the best way to propose changes to the codebase. We actively welcome your pull requests:
+
+1. Fork the repo and create your branch from `main`.
+2. If you've added code that should be tested, add tests.
+3. If you've changed APIs, update the documentation.
+4. Ensure the test suite passes.
+5. Make sure your code lints and is formatted with Black.
+6. Issue that pull request!
+
+## Issue Tracking
+
+We use GitHub issues to track public bugs and requests. Please ensure your description is clear and has sufficient instructions to be able to reproduce the issue.
+
+## Any questions?
+
+Don't hesitate to contact us if you have any questions. Contact [@lucasmiranda42](mailto:lucas_miranda@psych.mpg.de) or ask your question on issues.
+
+Thank you for your contributions!
