### 2023-05-03 df1709b
Fixed bug with embedding visualizations
