+++ deepof/visuals.py
@@ -739,7 +739,7 @@ def plot_transitions(
         soft_counts,
         breaks,
         exp_conditions,
-        bin_size=bin_size,
+        bin_size=(coordinates._frame_rate * bin_size if bin_size is not None else None),
         bin_index=bin_index,
         silence_diagonal=silence_diagonal,
         aggregate=(exp_conditions is not None),
@@ -980,54 +980,37 @@ def _filter_embeddings(
     soft_counts,
     breaks,
     supervised_annotations,
-    use_keys,
     exp_condition,
     bin_size,
     bin_index,
 ):
     """Auxiliary function to plot_embeddings. Filters all available data based on the provided keys and experimental condition."""
     # Get experimental conditions per video
+    if embeddings is None and supervised_annotations is None:
+        raise ValueError(
+            "Either embeddings, soft_counts, and breaks or supervised_annotations must be provided."
+        )
 
-    if exp_condition is None:
-        try:
+    try:
+        if exp_condition is None:
             exp_condition = list(embeddings._exp_conditions.values())[0].columns[0]
-        except AttributeError:
+
+        concat_hue = [
+            coordinates.get_exp_conditions[i][exp_condition].values[0]
+            for i in list(embeddings.keys())
+        ]
+        soft_counts = soft_counts.filter_videos(embeddings.keys())
+        breaks = breaks.filter_videos(embeddings.keys())
+
+    except AttributeError:
+        if exp_condition is None:
             exp_condition = list(supervised_annotations._exp_conditions.values())[
                 0
             ].columns[0]
 
-    concat_hue = [
-        i[exp_condition].values[0]
-        for i in list(
-            (
-                embeddings if embeddings is not None else supervised_annotations
-            )._exp_conditions.values()
-        )
-    ]
-    try:
-        soft_counts = soft_counts.filter_videos(embeddings.keys())
-        breaks = breaks.filter_videos(embeddings.keys())
-    except AttributeError:
-        pass
-
-    if use_keys is not None:
-        try:
-            try:
-                embeddings = {key: embeddings[key] for key in use_keys}
-                soft_counts = {key: soft_counts[key] for key in use_keys}
-                breaks = {key: breaks[key] for key in use_keys}
-            except AttributeError:
-                supervised_annotations = {
-                    key: supervised_annotations[key] for key in use_keys
-                }
-        except AttributeError:
-            raise ValueError(
-                "Either embeddings, soft_counts, and breaks or supervised_annotations must be provided."
-            )
         concat_hue = [
-            j
-            for i, j in zip(list(coordinates.get_exp_conditions.keys()), concat_hue)
-            if i in use_keys
+            coordinates.get_exp_conditions[i][exp_condition].values[0]
+            for i in list(supervised_annotations.keys())
         ]
 
     # Restrict embeddings, soft_counts and breaks to the selected time bin
@@ -1082,7 +1065,6 @@ def plot_normative_log_likelihood(
     ax: Any,
     add_stats: str,
     verbose: bool,
-    use_keys: list,
 ):
     """Plot a bar chart with normative log likelihoods per experimental condition, and compute statistics.
 
@@ -1094,7 +1076,6 @@ def plot_normative_log_likelihood(
         ax (plt.AxesSubplot): matplotlib axes where to render the plot
         add_stats (str): test to use. Mann-Whitney (non-parametric) by default. See statsannotations documentation for details.
         verbose (bool): if True, prints test results and p-value cutoffs. False by default.
-        use_keys (list): list of keys to use for plotting. If None (default), all keys are used.
 
     Returns:
         embedding_dataset (pd.DataFrame): embedding data frame with added normative scores per sample
[…]
