+++ pyproject.toml
@@ -45,6 +45,7 @@ tables = "^3.8.0"
 catboost = "^1.1.1"
 tornado = "6.1"
 jupyter-client = "7.3.2"
+pomegranate = "^1.0.0"
 
 [tool.poetry.group.dev.dependencies]
 black = "^22.8.0"

[…]
