### 2023-05-04 b6bdbad
Fixed paths in unsupervised tutorial, refactored tests
