+++ docs/source/deepof.annotation_utils.rst
@@ -1,5 +1,5 @@
 deepof.annotation_utils module
-=====================
+===============================
 
 .. automodule:: deepof.annotation_utils
    :members:
