+++ docs/source/deepof.model_utils.rst
@@ -1,5 +1,5 @@
 deepof.model_utils module
-=====================
+=========================
 
 .. automodule:: deepof.model_utils
    :members:
