### 2023-05-04 7afbd25
Fixed paths in unsupervised tutorial, refactored tests
