### 2023-05-04 8c3cad4
Fixed paths in unsupervised tutorial, refactored tests
