+++ requirements.txt
@@ -1,149 +1,139 @@
 absl-py==1.4.0
-aiofiles==22.1.0
-aiosqlite==0.19.0
 alabaster==0.7.13
 anyio==3.6.2
 appnope==0.1.3
-apricot-select==0.6.1
 argon2-cffi==21.3.0
 argon2-cffi-bindings==21.2.0
+arrow==1.2.3
 asttokens==2.2.1
 astunparse==1.6.3
 attrs==22.2.0
 av==10.0.0
-Babel==2.12.1
+Babel==2.11.0
 backcall==0.2.0
-beautifulsoup4==4.12.0
+beautifulsoup4==4.11.1
 black==22.12.0
-bleach==6.0.0
+bleach==5.0.1
 blosc2==2.0.0
-cachetools==5.3.0
+cachetools==5.2.1
 catboost==1.1.1
 certifi==2022.12.7
 cffi==1.15.1
-chardet==5.1.0
-charset-normalizer==3.1.0
+charset-normalizer==3.0.1
 click==8.1.3
 cloudpickle==2.2.1
-comm==0.1.3
+comm==0.1.2
 contourpy==1.0.7
 coverage==6.5.0
 cycler==0.11.0
 Cython==0.29.33
-dask==2023.3.2
+dask==2023.1.0
 dask-image==2022.9.0
-debugpy==1.6.6
+debugpy==1.6.5
 decorator==5.1.1
--e git+https://gitlab.mpcdf.mpg.de/lucasmir/deepof@b6bdbad5c5f0c30d7988c4b0429ce71c23566152#egg=deepof
 defusedxml==0.7.1
-deptry==0.8.0
 dm-tree==0.1.8
-docutils==0.18.1
+docutils==0.17.1
 entrypoints==0.4
 et-xmlfile==1.1.0
-exceptiongroup==1.1.1
+exceptiongroup==1.1.0
 executing==1.2.0
-fastjsonschema==2.16.3
-filelock==3.12.0
-flatbuffers==23.3.3
-fonttools==4.39.3
-fsspec==2023.3.0
+fastjsonschema==2.16.2
+flatbuffers==23.1.4
+fonttools==4.38.0
+fqdn==1.5.1
+fsspec==2022.11.0
 gast==0.4.0
-google-auth==2.17.0
+google-auth==2.16.0
 google-auth-oauthlib==0.4.6
 google-pasta==0.2.0
 graphviz==0.20.1
-grpcio==1.53.0
-h5py==3.8.0
-hypothesis==6.70.1
+grpcio==1.51.1
+h5py==3.7.0
+hypothesis==6.62.1
 idna==3.4
-imageio==2.27.0
+imageio==2.24.0
 imageio-ffmpeg==0.4.8
 imagesize==1.4.1
 imbalanced-learn==0.10.1
 imblearn==0.0
-importlib-metadata==6.1.0
-importlib-resources==5.12.0
+importlib-metadata==6.0.0
 iniconfig==2.0.0
-ipykernel==6.22.0
-ipython==8.11.0
+ipykernel==6.20.2
+ipython==8.8.0
 ipython-genutils==0.2.0
-ipywidgets==8.0.6
+ipywidgets==8.0.4
+isoduration==20.11.0
 jedi==0.18.2
 Jinja2==3.1.2
 joblib==1.2.0
-json5==0.9.11
+jsonpointer==2.3
 jsonschema==4.17.3
 jupyter==1.0.0
-jupyter-client==7.3.2
-jupyter-console==6.6.3
-jupyter-server==1.24.0
+jupyter-console==6.4.4
+jupyter-events==0.6.3
 jupyter-sphinx==0.4.0
-jupyter-ydoc==0.2.4
[…]
