### 2023-05-04 be923df
Fixed paths in unsupervised tutorial, refactored tests
