### 2023-05-05 fa26cf5
Fixed paths in unsupervised tutorial, refactored tests
