+++ docs/build/html/_sources/tutorial_notebooks/deepof_unsupervised_tutorial.ipynb.txt
@@ -37,7 +37,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 1,
+   "execution_count": null,
    "id": "d5b28605",
    "metadata": {},
    "outputs": [],
@@ -66,15 +66,7 @@
    "execution_count": 2,
    "id": "a2b46f72",
    "metadata": {},
-   "outputs": [
-    {
-     "name": "stderr",
-     "output_type": "stream",
-     "text": [
-      "OMP: Info #271: omp_set_nested routine deprecated, please use omp_set_max_active_levels instead.\n"
-     ]
-    }
-   ],
+   "outputs": [],
    "source": [
     "import copy\n",
     "import os\n",
@@ -136,20 +128,7 @@
    "execution_count": 4,
    "id": "1ce3f645",
    "metadata": {},
-   "outputs": [
-    {
-     "ename": "FileNotFoundError",
-     "evalue": "[Errno 2] No such file or directory: './tutorial_files/tutorial_project/Coordinates/deepof_coordinates.pkl'",
-     "output_type": "error",
-     "traceback": [
-      "\u001b[0;31m---------------------------------------------------------------------------\u001b[0m",
-      "\u001b[0;31mFileNotFoundError\u001b[0m                         Traceback (most recent call last)",
-      "Cell \u001b[0;32mIn[4], line 2\u001b[0m\n\u001b[1;32m      1\u001b[0m \u001b[38;5;66;03m# Load a previously saved project\u001b[39;00m\n\u001b[0;32m----> 2\u001b[0m my_deepof_project \u001b[38;5;241m=\u001b[39m \u001b[43mdeepof\u001b[49m\u001b[38;5;241;43m.\u001b[39;49m\u001b[43mdata\u001b[49m\u001b[38;5;241;43m.\u001b[39;49m\u001b[43mload_project\u001b[49m\u001b[43m(\u001b[49m\u001b[38;5;124;43m\"\u001b[39;49m\u001b[38;5;124;43m./tutorial_files/tutorial_project\u001b[39;49m\u001b[38;5;124;43m\"\u001b[39;49m\u001b[43m)\u001b[49m\n",
[…]
