### 2023-05-05 1039cfa
Fixed paths in unsupervised tutorial, refactored tests
