### 2023-05-05 aaa2544
Added option to edit arena parameters after project creation
