### 2023-05-06 eb95c5b
Fixed paths in unsupervised tutorial, refactored tests
