### 2023-05-08 5d7bd31
Refactored unsupervised training script
