### 2023-05-08 16cf4fb
Refactored and fixed visualization bugs
