+++ deepof/visuals.py
@@ -461,7 +461,7 @@ def plot_gantt(
         behavior_ticks = np.array(
             [
                 f"Cluster {cluster}"
-                for cluster in range(gantt.shape[0])
+                for cluster in range(n_features)
                 if behaviors_to_plot is None or cluster in behaviors_to_plot
             ]
         )
@@ -471,6 +471,7 @@ def plot_gantt(
         )
 
     plt.xticks([])
+
     plt.yticks(
         np.array(range(gantt.shape[0])) + 0.5,
         # Concatenate cluster IDs and checkpoint names if they exist
@@ -527,7 +528,7 @@ def plot_enrichment(
     # Time selection parameters
     bin_size: int = None,
     bin_index: int = 0,
-    precomputed: np.ndarray = None,
+    precomputed_bins: np.ndarray = None,
     # Visualization parameters
     exp_condition: str = None,
     exp_condition_order: list = None,
@@ -550,7 +551,7 @@ def plot_enrichment(
         min_confidence (float): minimum confidence in cluster assignments used for quality control filtering.
         bin_size (int): bin size for time filtering.
         bin_index (int): index of the bin of size bin_size to select along the time dimension.
-        precomputed (np.ndarray): precomputed time bins. If provided, bin_size and bin_index are ignored.
+        precomputed_bins (np.ndarray): precomputed time bins. If provided, bin_size and bin_index are ignored.
         add_stats (str): test to use. Mann-Whitney (non-parametric) by default. See statsannotations documentation for details.
         verbose (bool): if True, prints test results and p-value cutoffs. False by default.
         ax (plt.AxesSubplot): axes where to plot the current figure. If not provided, new figure will be created.
@@ -591,7 +592,7 @@ def plot_enrichment(
         exp_conditions=exp_conditions,
         bin_size=(coordinates._frame_rate * bin_size if bin_size is not None else None),
         bin_index=bin_index,
-        precomputed=precomputed,
+        precomputed=precomputed_bins,
         normalize=normalize,
     )
 
@@ -647,6 +648,14 @@ def plot_enrichment(
             .tolist()
         ]
         pairs = [item for sublist in pairs for item in sublist]
+
+        # Remove elements from pairs if clusters are not present in the enrichment data frame
+        pairs = [
+            p
+            for p in pairs
+            if p[0][0] in enrichment["cluster"].values
+            and p[1][0] in enrichment["cluster"].values
+        ]
 
         annotator = Annotator(
             ax,
@@ -984,6 +993,7 @@ def _filter_embeddings(
     exp_condition,
     bin_size,
     bin_index,
+    precomputed_bins,
 ):
     """Auxiliary function to plot_embeddings. Filters all available data based on the provided keys and experimental condition."""
     # Get experimental conditions per video
@@ -1015,7 +1025,15 @@ def _filter_embeddings(
         ]
 
     # Restrict embeddings, soft_counts and breaks to the selected time bin
-    if bin_size is not None:
+    if precomputed_bins is not None:
+        embeddings, soft_counts, breaks, _ = deepof.post_hoc.select_time_bin(
+            embeddings,
+            soft_counts,
+            breaks,
+            precomputed=(precomputed_bins == bin_index),
+        )
+
+    elif bin_size is not None:
         if embeddings is not None:
             embeddings, soft_counts, breaks, _ = deepof.post_hoc.select_time_bin(
                 embeddings,
@@ -1059,7 +1077,7 @@ def _filter_embeddings(
 
 
 def plot_normative_log_likelihood(
-    coordinates: coordinates,
+    embeddings: table_dict,
     exp_condition: str,
     embedding_dataset: pd.DataFrame,
     normative_model: str,
@@ -1070,7 +1088,7 @@ def plot_normative_log_likelihood(
[…]
