+++ deepof/data.py
@@ -1261,6 +1261,16 @@ class Coordinates:
             self._scales[vid_idcs[vid_idx]] = edited_scales[vid_idx]
             self._arena_params[vid_idcs[vid_idx]] = edited_arena_params[vid_idx]
 
+        self.save(
+            filename=os.path.join(
+                self._project_path,
+                self._project_name,
+                "Coordinates",
+                "deepof_coordinates",
+            ),
+            timestamp=False,
+        )
+
     def save(self, filename: str = None, timestamp: bool = True):
         """Save the current state of the Coordinates object to a pickled file.
 
