### 2023-05-11 87562ab
Refactored and fixed visualization bugs
