### 2023-05-16 8b88938
Refactored and fixed visualization bugs
