+++ deepof/utils.py
@@ -3,7 +3,7 @@
 # module deepof
 
 """Functions and general utilities for the deepof package."""
-
+import copy
 from copy import deepcopy
 from dask_image.imread import imread
 from itertools import combinations, product
@@ -12,6 +12,8 @@ from scipy.signal import savgol_filter
 from shapely.geometry import Polygon
 from sklearn import mixture
 from sklearn.feature_selection import VarianceThreshold
+from sklearn.experimental import enable_iterative_imputer  # noqa
+from sklearn.impute import IterativeImputer
 from sklearn.preprocessing import StandardScaler, MinMaxScaler, RobustScaler
 from tqdm import tqdm
 from typing import Tuple, Any, List, Union, NewType
@@ -26,6 +28,9 @@ import pandas as pd
 import regex as re
 import ruptures as rpt
 import tensorflow as tf
+import warnings
+
+import deepof.data
 
 # DEFINE CUSTOM ANNOTATED TYPES #
 project = NewType("deepof_project", Any)
@@ -36,7 +41,9 @@ table_dict = NewType("deepof_table_dict", Any)
 # CONNECTIVITY AND GRAPH REPRESENTATIONS
 
 
-def connect_mouse_topview(animal_ids=None, exclude_bodyparts: list = None) -> nx.Graph:
+def connect_mouse(
+    animal_ids=None, exclude_bodyparts: list = None, graph_preset: str = "deepof_14"
+) -> nx.Graph:
     """Create a nx.Graph object with the connectivity of the bodyparts in the DLC topview model for a single mouse.
 
     Used later for angle computing, among others.
@@ -44,6 +51,7 @@ def connect_mouse_topview(animal_ids=None, exclude_bodyparts: list = None) -> nx
     Args:
         animal_ids (str): if more than one animal is tagged, specify the animal identyfier as a string.
         exclude_bodyparts (list): Remove the specified nodes from the graph.
+        graph_preset (str): Connectivity preset to use. Currently supported: "deepof_14" and "deepof_8".
 
     Returns:
         connectivity (nx.Graph)
@@ -57,16 +65,29 @@ def connect_mouse_topview(animal_ids=None, exclude_bodyparts: list = None) -> nx
     connectivities = []
 
     for animal_id in animal_ids:
-        connectivity = {
-            "Nose": ["Left_ear", "Right_ear"],
-            "Spine_1": ["Center", "Left_ear", "Right_ear"],
-            "Center": ["Left_fhip", "Right_fhip", "Spine_2"],
-            "Spine_2": ["Left_bhip", "Right_bhip", "Tail_base"],
-            "Tail_base": ["Tail_1"],
-            "Tail_1": ["Tail_2"],
-            "Tail_2": ["Tail_tip"],
+        connectivity_dict = {
+            "deepof_14": {
+                "Nose": ["Left_ear", "Right_ear"],
+                "Spine_1": ["Center", "Left_ear", "Right_ear"],
+                "Center": ["Left_fhip", "Right_fhip", "Spine_2"],
+                "Spine_2": ["Left_bhip", "Right_bhip", "Tail_base"],
+                "Tail_base": ["Tail_1"],
+                "Tail_1": ["Tail_2"],
+                "Tail_2": ["Tail_tip"],
+            },
+            "deepof_8": {
+                "Nose": ["Left_ear", "Right_ear"],
+                "Center": [
+                    "Left_fhip",
+                    "Right_fhip",
+                    "Tail_base",
+                    "Left_ear",
+                    "Right_ear",
+                ],
+                "Tail_base": ["Tail_tip"],
+            },
         }
-        connectivity = nx.Graph(connectivity)
[…]
