+++ deepof/data.py
@@ -21,8 +21,6 @@ from shapely.geometry import Polygon
 from shutil import rmtree
 from sklearn import random_projection
 from sklearn.decomposition import KernelPCA
-from sklearn.experimental import enable_iterative_imputer  # noqa
-from sklearn.impute import IterativeImputer
 from sklearn.manifold import TSNE
 from sklearn.preprocessing import (
     MinMaxScaler,
@@ -95,6 +93,7 @@ class Project:
         self,
         animal_ids: List = [""],
         arena: str = "polygonal-manual",
+        bodypart_graph: str = "deepof_14",
         enable_iterative_imputation: bool = 250,
         exclude_bodyparts: List = tuple([""]),
         exp_conditions: dict = None,
@@ -117,6 +116,7 @@ class Project:
         Args:
             animal_ids (list): list of animal ids.
             arena (str): arena type. Can be one of "circular-autodetect", "circular-manual", or "polygon-manual".
+            bodypart_graph (str): body part scheme to use for the analysis. Defaults to None, in which case the program will attempt to select it automatically based on the available body parts.
             enable_iterative_imputation (bool): whether to use iterative imputation for occluded body parts. Recommended if several animals are present, but slower.
             exclude_bodyparts (list): list of bodyparts to exclude from analysis.
             exp_conditions (dict): dictionary with experiment IDs as keys and experimental conditions as values.
@@ -177,6 +177,7 @@ class Project:
         self.angles = True
         self.animal_ids = animal_ids
         self.areas = True
+        self.bodypart_graph = bodypart_graph
         self.connectivity = None
         self.distances = "all"
         self.ego = False
@@ -383,8 +384,10 @@ class Project:
 
         # Update body part connectivity graph, taking detected or specified body parts into account
         model_dict = {
-            "{}mouse_topview".format(aid): deepof.utils.connect_mouse_topview(
-                aid, exclude_bodyparts=self.exclude_bodyparts
+            "{}mouse_topview".format(aid): deepof.utils.connect_mouse(
+                aid,
+                exclude_bodyparts=self.exclude_bodyparts,
+                graph_preset=self.bodypart_graph,
             )
             for aid in self.animal_ids
         }
@@ -419,6 +422,8 @@ class Project:
 
             tab_dict[key] = pd.concat([x, y], axis=1).sort_index(axis=1)
             lik_dict[key] = lik.fillna(0.0)
+
+        lik_dict = TableDict(lik_dict, typ="quality", animal_ids=self.animal_ids)
 
         if self.smooth_alpha:
 
@@ -467,30 +472,10 @@ class Project:
             if verbose:
                 print("Iterative imputation of ocluded bodyparts...")
 
-            for k, tab in tab_dict.items():
+            tab_dict = deepof.utils.iterative_imputation(self, tab_dict, lik_dict)
 
-                scaler = StandardScaler()
-                imputed = IterativeImputer(
-                    skip_complete=True,
-                    max_iter=self.enable_iterative_imputation,
-                    n_nearest_features=tab.shape[1] // len(self.animal_ids) - 1,
-                    tol=1e-1,
-                ).fit_transform(scaler.fit_transform(tab))
-                imputed = pd.DataFrame(
-                    scaler.inverse_transform(imputed),
-                    index=tab.index,
-                    columns=tab.loc[:, tab.isnull().mean(axis=0) != 1.0].columns,
-                )
-
-                tab.update(imputed)
-                tab_dict[k] = tab
-
-                if tab.shape != imputed.shape:
-                    warnings.warn(
-                        "Some of the body parts have zero measurements. Iterative imputation skips these,"
-                        " which could bring problems downstream. A possible solution could be to refine "
-                        "DLC tracklets."
-                    )
[…]
