### 2023-05-17 bb9e40b
Refactored and fixed visualization bugs
