### 2023-05-17 b0adf48
Refactored and fixed visualization bugs
