### 2023-05-18 ea9aed3
Refactored and fixed visualization bugs
