### 2023-05-18 26b8ed4
Refactored and fixed visualization bugs
