### 2023-05-18 fd84ad2
Refactored and fixed visualization bugs
