### 2023-05-18 6e7cb53
Updated documentation and refactored
