+++ deepof/data.py
@@ -15,7 +15,6 @@ For a detailed tutorial on how to use this module, see the advanced tutorials in
 
 
 from collections import defaultdict
-from joblib import delayed, Parallel, parallel_backend
 from pkg_resources import resource_filename
 from shapely.geometry import Polygon
 from shutil import rmtree
@@ -31,7 +30,7 @@ from sklearn.preprocessing import (
 from time import time
 from tqdm import tqdm
 from typing import Any, NewType, Union
-from typing import Dict, List, Tuple
+from typing import Dict, List, Tuple, Any
 import copy
 import datetime
 import math
@@ -91,7 +90,7 @@ class Project:
 
     def __init__(
         self,
-        animal_ids: List = [""],
+        animal_ids: List = None,
         arena: str = "polygonal-manual",
         bodypart_graph: str = "deepof_14",
         enable_iterative_imputation: bool = 250,
@@ -103,7 +102,7 @@ class Project:
         likelihood_tol: float = 0.75,
         model: str = "mouse_topview",
         project_name: str = "deepof_project",
-        project_path: str = deepof.utils.os.path.join("."),
+        project_path: str = os.path.join("."),
         video_path: str = None,
         table_path: str = None,
         smooth_alpha: float = 1,
@@ -153,14 +152,14 @@ class Project:
         self.videos = sorted(
             [
                 vid
-                for vid in deepof.utils.os.listdir(self.video_path)
+                for vid in os.listdir(self.video_path)
                 if vid.endswith(video_format) and not vid.startswith(".")
             ]
         )
         self.tables = sorted(
             [
                 tab
-                for tab in deepof.utils.os.listdir(self.table_path)
+                for tab in os.listdir(self.table_path)
                 if tab.endswith(self.table_format) and not tab.startswith(".")
             ]
         )
@@ -175,7 +174,7 @@ class Project:
 
         # Set the rest of the init parameters
         self.angles = True
-        self.animal_ids = animal_ids
+        self.animal_ids = animal_ids if animal_ids is not None else [""]
         self.areas = True
         self.bodypart_graph = bodypart_graph
         self.connectivity = None
@@ -312,7 +311,7 @@ class Project:
             self.videos,
         )
 
-    def load_tables(self, verbose: bool = True) -> deepof.utils.Tuple:
+    def load_tables(self, verbose: bool = True) -> Tuple:
         """Load videos and tables into dictionaries.
 
         Args:
@@ -338,7 +337,7 @@ class Project:
             tab_dict = {}
             for tab in self.tables:
                 loaded_tab = pd.read_hdf(
-                    deepof.utils.os.path.join(self.table_path, tab), dtype=float
+                    os.path.join(self.table_path, tab), dtype=float
                 )
 
                 # Adapt index to be compatible with downstream processing
@@ -352,7 +351,7 @@ class Project:
 
             tab_dict = {
                 deepof.utils.re.findall("(.*?)DLC", tab)[0]: pd.read_csv(
-                    deepof.utils.os.path.join(self.table_path, tab),
+                    os.path.join(self.table_path, tab),
                     index_col=0,
                     low_memory=False,
                 )
@@ -595,15 +594,17 @@ class Project:
 
             exp_table = pd.DataFrame()
 
-            for id in self.animal_ids:
+            for aid in self.animal_ids:
 
-                if id == "":
-                    id = None
+                if aid == "":
+                    aid = None
 
                 # get the current table for the current animal
-                current_table = tab.loc[:, deepof.utils.filter_columns(tab.columns, id)]
+                current_table = tab.loc[
+                    :, deepof.utils.filter_columns(tab.columns, aid)
+                ]
                 current_table = current_table.apply(
-                    lambda x: deepof.utils.compute_areas(x, animal_id=id), axis=1
[…]
