+++ .gitlab-ci.yml
@@ -15,7 +15,7 @@ tutorials:
     - pip install -U --no-cache-dir gdown --pre
     - pip install -q -e . --progress-bar off
     - python -m ipykernel install --user --name deepof --display-name "Python (deepof)"
-    - gdown -q --folder https://drive.google.com/drive/folders/1DBK-z5KIIzhpF0l-M1Shgx3kLUwpENWc?usp=sharing
+    - gdown -q --folder "https://drive.google.com/drive/folders/1DBK-z5KIIzhpF0l-M1Shgx3kLUwpENWc?usp=sharing"
 
   script:
     - papermill docs/source/tutorial_notebooks/deepof_preprocessing_tutorial.ipynb preprocessing_output.ipynb
