### 2023-05-19 mlfpm/deepof#8 @edeno
[JOSS] Author contributions
Regarding contribution and authorship of the JOSS paper, it seems like the first author @lucasmiranda42 has made substantial contributions to the software package, but I am unclear on the contributions of the other authors. Could you list the contributions of the other authors here?

xref: https://github.com/openjournals/joss-reviews/issues/5394
