### 2023-05-19 mlfpm/deepof#10 @edeno
[JOSS] Poetry installation failure
I also tried installing via poetry and received the following error:
```bash
poetry install

Poetry could not find a pyproject.toml file in /home/edeno/Documents/Github/deepof or its parents
```

xref: https://github.com/openjournals/joss-reviews/issues/5394
