### 2023-05-19 mlfpm/deepof#10 @lucasmiranda42
Dear @edeno, first of all thank you so much for reviewing DeepOF!

I start with this one because it seems like the easiest to troubleshoot. Sorry to ask, but are you in the correct directory? is pyproject.toml in the directory from which you're calling `poetry install`?

I realize the documentation may be a bit ambiguous here; let me know if you manage and I'll edit the text to make it clearer
