### 2023-05-19 mlfpm/deepof#10 @edeno
Ah I actually discovered the problem. I didn't have git-lfs installed, which I think caused an error in cloning. If you could include this in your documentation that would be great.

Another small issue discovered. Python <3.11 is required.
