### 2023-05-19 mlfpm/deepof#8 @lucasmiranda42
Absolutely. Joeri Bordes and Mathias Schmidt contributed with the overall design and testing of the API and scope of the package. They also regularly test new features with different datasets before any new release. Benno Pütz and Bertram Müller-Myhsok contributed with knowledge on backend design, mostly. They aided substantially with developing the CI pipeline, interleaving of the different classes, and ideas behind many of the models implemented in `deepof.models`, even when I was mostly behind the keyboard.

I am happy to clarify and/or discuss in more detail, but I think all of them deserve to be in the author's list if you agree :)
