### 2023-05-19 mlfpm/deepof#10 @lucasmiranda42
Wonderful, glad you managed! I'll add the Python version to the documentation.
