### 2023-05-19 mlfpm/deepof#9 @lucasmiranda42
Hmmm, interesting. Our CI pipeline implements a ['deploy' job](https://gitlab.mpcdf.mpg.de/lucasmir/deepof/-/jobs/2172045) testing exactly this, but only using the python:3.9.14 Docker image that runs on Debian. I'll add testing pipelines in other OS and troubleshoot
