### 2023-05-19 mlfpm/deepof#11 @edeno
[JOSS] Docker installation
It would be great to document what a user should do after running the docker image. For example, I tried:
```bash
root@05b17dd067c8:/# ipython
Python 3.9.14 (main, Oct  5 2022, 15:22:07)
Type 'copyright', 'credits' or 'license' for more information
IPython 8.11.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: import deepof
Cell In[1], line 1
----> 1 import deepof

ModuleNotFoundError: No module named 'deepof'
```

xref: https://github.com/openjournals/joss-reviews/issues/5394
