### 2023-05-19 mlfpm/deepof#11 @lucasmiranda42
Thank you for trying all installation methods! Indeed, one must install DeepOF after loading the docker image (either using pip or from the repo). Would you recommend editing the documentation, or adding DeepOF to the Docker image directly?
