### 2023-05-19 mlfpm/deepof#11 @edeno
I would add DeepOF to the docker image directly. Might as well cut down on an additional step for the user.
