### 2023-05-20 mlfpm/deepof#11 @lucasmiranda42
Dear @edeno,

I just pushed an updated Docker image with DeepOF loaded onto it, which should address the problem presented above. I'll close this issue for now, but feel free to reopen if needed, of course!
