### 2023-05-20 mlfpm/deepof#9 @lucasmiranda42
Dear @edeno,

I just pushed a new version to PyPI, with a much less restricted set of requirements that works fine across different OS (Ubuntu included). I tested it with Python 3.10.6 in Ubuntu 23.10.

Feel free to retry `pip install` (preferable in a new environment) and let me know if the issue is solved. I'll close it for now!
