### 2023-05-20 8c53d0a
Released 0.3 version, fixed requirements.txt, and updated Dockerfile to include DeepOF installation
