+++ requirements.txt
@@ -1,224 +1,37 @@
-absl-py==1.4.0
-alabaster==0.7.13
-anyio==3.6.2
-appnope==0.1.3
-argon2-cffi==21.3.0
-argon2-cffi-bindings==21.2.0
-arrow==1.2.3
-asttokens==2.2.1
-astunparse==1.6.3
-attrs==22.2.0
-av==10.0.0
-Babel==2.11.0
-backcall==0.2.0
-beautifulsoup4==4.11.1
-black==22.12.0
-bleach==5.0.1
-blosc2==2.0.0
-cachetools==5.2.1
-catboost==1.1.1
-certifi==2022.12.7
-cffi==1.15.1
-charset-normalizer==3.0.1
-click==8.1.3
-cloudpickle==2.2.1
-comm==0.1.2
-contourpy==1.0.7
-coverage==6.5.0
-cycler==0.11.0
-Cython==0.29.33
-dask==2023.1.0
-dask-image==2022.9.0
-debugpy==1.6.5
-decorator==5.1.1
-defusedxml==0.7.1
-dm-tree==0.1.8
-docutils==0.17.1
-entrypoints==0.4
-et-xmlfile==1.1.0
-exceptiongroup==1.1.0
-executing==1.2.0
-fastjsonschema==2.16.2
-flatbuffers==23.1.4
-fonttools==4.38.0
-fqdn==1.5.1
-fsspec==2022.11.0
-gast==0.4.0
-google-auth==2.16.0
-google-auth-oauthlib==0.4.6
-google-pasta==0.2.0
-graphviz==0.20.1
-grpcio==1.51.1
-h5py==3.7.0
-hypothesis==6.62.1
-idna==3.4
-imageio==2.24.0
-imageio-ffmpeg==0.4.8
-imagesize==1.4.1
-imbalanced-learn==0.10.1
-imblearn==0.0
-importlib-metadata==6.0.0
-iniconfig==2.0.0
-ipykernel==6.20.2
-ipython==8.8.0
-ipython-genutils==0.2.0
-ipywidgets==8.0.4
-isoduration==20.11.0
-jedi==0.18.2
-Jinja2==3.1.2
-joblib==1.2.0
-jsonpointer==2.3
-jsonschema==4.17.3
-jupyter==1.0.0
-jupyter-console==6.4.4
-jupyter-events==0.6.3
-jupyter-sphinx==0.4.0
-jupyter_client==7.4.9
-jupyter_core==5.1.3
-jupyter_server==2.1.0
-jupyter_server_terminals==0.4.4
-jupyterlab-pygments==0.2.2
-jupyterlab-widgets==3.0.5
-keras==2.11.0
-keras-tcn==3.5.0;platform_machine != 'arm64' or platform_system != 'Darwin'
-keras-tcn-macos==1.0;platform_machine == 'arm64' and platform_system == 'Darwin'
-keras-tuner==1.1.3
-kiwisolver==1.4.4
-kt-legacy==1.0.4
-libclang==15.0.6.1
-llvmlite==0.39.1
-locket==1.0.0
-Markdown==3.4.1
-MarkupSafe==2.1.2
-matplotlib==3.6.3
-matplotlib-inline==0.1.6
-mistune==2.0.4
-msgpack==1.0.4
-mypy-extensions==0.4.3
-nbclassic==0.4.8
-nbclient==0.7.2
-nbconvert==7.2.8
-nbformat==5.7.3
-nbsphinx==0.8.11
-nest-asyncio==1.5.6
-networkx==2.8.8
-notebook==6.5.2
-notebook_shim==0.2.2
-numba==0.56.4
-numexpr==2.8.4
-numpy==1.23.5
-oauthlib==3.2.2
[…]
