### 2023-05-20 mlfpm/deepof#10 @lucasmiranda42
Dear @edeno, I removed the git lfs requirement from the repo and updated the documentation to specify the Python version. I'll close the issue for now; let me know if you would like me to implement any further changes :)
