### 2023-05-20 6843396
Released 0.3 version, fixed requirements.txt, and updated Dockerfile to include DeepOF installation
