+++ docs/build/html/_sources/tutorial_notebooks/deepof_supervised_tutorial.ipynb.txt
@@ -39,18 +39,32 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 1,
    "id": "764990b0",
    "metadata": {},
    "outputs": [],
    "source": [
-    "# # If using Google colab, uncomment and run this cell to set up the environment\n",
+    "# # If using Google colab, uncomment and run this cell and the one below to set up the environment\n",
+    "# # Note: becasuse of how colab handles the installation of local packages, this cell will kill your runtime.\n",
+    "# # This is not an error! Just continue with the cells below.\n",
+    "# import os\n",
     "# !git clone -q https://github.com/mlfpm/deepof.git\n",
     "# !pip install -q -e deepof --progress-bar off\n",
-    "# !gdown -q --folder https://drive.google.com/drive/folders/1DBK-z5KIIzhpF0l-M1Shgx3kLUwpENWc?usp=sharing\n",
+    "# os.chdir(\"deepof\")\n",
+    "# !gdown -q --folder \"https://drive.google.com/drive/folders/1DBK-z5KIIzhpF0l-M1Shgx3kLUwpENWc?usp=sharing\"\n",
+    "# os.kill(os.getpid(), 9)"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "id": "50799c68",
+   "metadata": {},
+   "outputs": [],
+   "source": [
     "# import os, warnings\n",
     "# warnings.filterwarnings('ignore')\n",
-    "# os.kill(os.getpid(), 9)"
+    "# os.chdir(\"deepof\")"
    ]
   },
   {
