+++ requirements.txt
@@ -36,3 +36,4 @@ tqdm>=4.64.0
 tqdm>=4.65.0
 umap_learn>=0.5.3
 jupyter_sphinx>=0.4.0
+nbsphinx>=0.8.7
