### 2023-05-22 mlfpm/deepof#11 @edeno
Okay I was able to import deepof successfully in this docker container. One question for you is would it also be beneficial to have deeplabcut in the container as well since the two libraries are dependent or is that out of scope in your mind?
