### 2023-05-22 mlfpm/deepof#9 @edeno
Also confirming pip installation was successful.
