### 2023-05-23 mlfpm/deepof#12 @edeno
[JOSS] Figure summarizing functionality
This is not necessary, but in your documentation after this line:
> You can use this package to either extract pre-defined motifs from the time series (such as time-in-zone, climbing, basic social interactions) or to embed your data into a sequence-aware latent space to extract meaningful motifs in an unsupervised way! Both of these can be used within the package, for example, to automatically compare user-defined experimental groups.

It would be nice for the user to see a basic figure illustrating what your package can do. It would help users more quickly understand what your package can do. You more or less have them in your documentation but you have to dig for them for users to see it.

xref: https://github.com/openjournals/joss-reviews/issues/5394
