### 2023-05-23 mlfpm/deepof#13 @edeno
[JOSS] Statement of Need - differences from existing software
In the manuscript, the authors state that there exist software packages that can detect behavior in both supervised and unsupervised manners. The authors then go onto state that what distinguishes *their software package* is supervised and unsupervised pipelines. Can the authors further clarify the differences here? Is it the second part the sentence that is the distinguishing factor? Are the supervised and unsupervised models different? It might be nice to have more than one sentence clarifying these differences.

xref: https://github.com/openjournals/joss-reviews/issues/5394
