### 2023-05-23 mlfpm/deepof#12 @lucasmiranda42
Dear @edeno, thanks! I have added a simple figure describing what the package does after the mentioned line. I think it's a pretty cool suggestion! Let me know what you think or if you'd suggest any modifications.

Best!
Lucas
