### 2023-05-23 c617d56
Added figure to landing page of documentation
