### 2023-05-24 mlfpm/deepof#13 @lucasmiranda42
Dear @edeno,

I have modified the text to further clarify the additions to the state of the art, and added the figure requested in the other issue to the manuscript, as I felt it could help with the description of the workflow here too.

Best, and let me know what you think!
Lucas
