### 2023-05-26 mlfpm/deepof#16 @edeno
[JOSS] Tutorial dependencies
Running through the data formatting [tutorial](https://deepof.readthedocs.io/en/latest/tutorial_notebooks/deepof_preprocessing_tutorial.html), I needed to install imageio-ffmpeg, pytables

xref: https://github.com/openjournals/joss-reviews/issues/5394
