### 2023-05-27 mlfpm/deepof#16 @lucasmiranda42
Hmmm. I'll have a look! Which installation did you follow? My bet is that the pip requirements I created for 0.3 are missing some optional dependencies of the parent packages, which are not optional to DeepOF
