### 2023-05-27 mlfpm/deepof#16 @edeno
This was the pip install.
