### 2023-05-29 mlfpm/deepof#16 @lucasmiranda42
Dear @edeno,

The latest release (deepof==0.4.5) should fix these issues. The main modification is the requirements.txt file, as stated. Please reopen and let me know if you still have problems!

Best, and thank you once again,
Lucas
