+++ .gitlab-ci.yml
@@ -36,10 +36,10 @@ tutorials:
 
   before_script:
     - pip install papermill
-    - pip install -U --no-cache-dir gdown --pre
     - pip install -q -e . --progress-bar off
     - python -m ipykernel install --user --name deepof --display-name "Python (deepof)"
-    - gdown -q --folder "https://drive.google.com/drive/folders/1DBK-z5KIIzhpF0l-M1Shgx3kLUwpENWc?usp=sharing"
+    - curl --output tutorial_files.zip https://datashare.mpcdf.mpg.de/s/knF7t78isQuIAr0/download
+    - unzip tutorial_files.zip
 
   script:
     - papermill docs/source/tutorial_notebooks/deepof_preprocessing_tutorial.ipynb preprocessing_output.ipynb
