+++ .gitlab-ci.yml
@@ -35,6 +35,8 @@ tutorials:
     entrypoint: [ "" ]
 
   before_script:
+    - apt update
+    - apt install -y curl unzip
     - pip install papermill
     - pip install -q -e . --progress-bar off
     - python -m ipykernel install --user --name deepof --display-name "Python (deepof)"
