### 2023-06-01 #5394 @emdupre
Thank you for generating those proofs, @lucasmiranda42, and for your engagement in the process so far !

It looks like all issues from the initial round of review have now been closed. @edeno and @cellistigs, would you be able to confirm if these changes have addressed your previous concerns ?
🙏  Thank you both very much for your thoughtful comments to date !
