### 2023-06-02 mlfpm/deepof#21 @edeno
[JOSS] Generality of this software
One concern I have overall is the generality of this software. In many places the body parts are hard coded and assume a top-down camera view. I know you've mentioned it a bit in the manuscript, but can you elaborate more on how you expect this to be useful to the general community given these assumptions. It would be useful if the software could at least have the potential of generalization in some of these cases.

xref: https://github.com/openjournals/joss-reviews/issues/5394
