### 2023-06-03 mlfpm/deepof#21 @lucasmiranda42
Dear @edeno,

Than you very much for your question! This is indeed an important point. The idea behind the scheme of body parts that are required is to provide annotation tools for the most common scenarios out of the box, but I agree with you that this leads to a decrease in generality.

There are two main places where this is a problem, that could potentially be extended in the future in different ways. These are:

1) The supervised annotation pipeline, where rules and pretrained models need specific body parts to run. Here, the current implementation will skip annotations for which labels are missing (you'll just get a NaN label), but no errors should be raised. This prevents the users from having to skip the entire pipeline if just one or a few labels are missing. Extending this pipeline with alternative labels could add flexibility, but I strongly feel is out of scope at the moment. This workflow aims to be a quick annotation tool for one of the most common scenarios, "trading off ease of use for flexibility", as we put in the paper.

2) In the unsupervised pipeline, there is the option to use a graph representation that, at the moment, uses a given set of nodes. Extensions to this in particular are indeed of interest at the moment, and new labelling schemes are starting to be supported (see [this issue](https://github.com/mlfpm/deepof/issues/7)). As I agree this is a pressing issue that would potentially lead to the package being used by more people, I'm adding a new tutorial on custom labelling schemes for the unsupervised pipeline. I'll comment again when ready :)

Best, and thank you once again for taking the time to review the package!
Lucas
