### 2023-06-03 mlfpm/deepof#21 @lucasmiranda42
The documentation has been modified to mention both labelling schemes supported out of the box, and a new tutorial has been added to explain how to proceed with a custom labelling scheme that does not fit with either.

I hope this addresses the issue! I'm closing it for now, but let me know if you still have concerns :)
Best, and thank you once again!
