### 2023-06-07 #5394 @emdupre
Thank you, @lucasmiranda42 !

As a follow-up, could you please :
- [ ] Change the title of the Zenodo archive to match the paper; that is, it should be "DeepOF: a Python package for supervised and unsupervised pattern recognition in mice motion tracking data". You can change this without creating a new record by editing the metadata as [described here](https://help.zenodo.org/#technical)

I will also have comments for you on the paper shortly !
