### 2023-06-09 #5394 @emdupre
Thank you, @lucasmiranda42 ! I have finished reviewing the paper and three editorial requests :

- [ ] On L27, please move the citation for Bourdes et al., (2022) to the end of the first clause, as "As demonstrated in the accompanying scientific application paper by Bourdes et al., (2022), ...." for clarity. You can use the `@author:2001` format for appropriate in-text rendering.
- [ ] The `title` field of [`Nilsson2020`](https://github.com/mlfpm/deepof/blob/e1dcfee1d81532e38303e84389471da184fdcb07/paper/paper.bib#L52) should be in double-brackets to ensure correct capitalization ; more generally, please review the titles of references to ensure correct in-text formatting.
- [ ] For the author affiliations, please remove the associated postal code (i.e., 80804).
