### 2023-06-09 #5394 @emdupre
Thank you, @lucasmiranda42 !

I'm now happy to recommend _DeepOF_ for publication 🚀 I'll now generate the a finalized proof ; please read through and review it for any last minute corrections.
Congratulations on this work, and thank you to @cellistigs and @edeno for your time and expertise in reviewing this submission !
