### 2023-06-09 #5394 @lucasmiranda42
One small detail is that the DOI appears as not found when I click on it, but I reckon that's by design and will change upon acceptance, right? :)
