### 2023-06-09 #5394 @emdupre
Yes exactly ; the DOI will not be finalized until publication 👍  But thank you for confirming your review of the proof !
