### 2023-06-12 #5394 @Kevin-Mattheus-Moerman
@emdupre good job editing this one :) there seems to be no changes required.

@lucasmiranda42 I have check your repository, this review, the archive, and the paper and all seems in order. I'll now proceed to process this for acceptance in JOSS.
