### 2023-06-20 #5394 @lucasmiranda42
Dear @emdupre, @Kevin-Mattheus-Moerman,

Thank you once again for your help during the revision process! I write because I'm including this paper as part of my doctoral thesis, and I've been asked to support it with an explicit permission from the publisher for the article to be included. Do you happen to know how to proceed in these cases?

Best, and thank you very much!
Lucas
