### 2023-06-20 #5394 @Kevin-Mattheus-Moerman
@lucasmiranda42 JOSS papers are fully open access. Copyright of JOSS papers is retained by submitting authors and accepted papers are subject to a Creative Commons Attribution 4.0 International License. Hence you are permitted to use the paper in your doctoral thesis as you see fit. Hope this addresses your query.
